+++ CHANGELOG.md
@@ -16,6 +16,11 @@
 
 * `DockerPlatform`: Remove duplicate auto-mounts (#257).
 
+## DEPRECATION
+
+* `DockerRequirements`: The `resources:` setting has been deprecated and will be removed in Viash 0.7.0. Please use `copy:` instead.
+
+* `DockerRequirements`: The `privileged:` setting has been deprecated and will be removed in Viash 0.7.0. Please use `run_args: "--privileged"` instead.
 
 # Viash 0.6.2